# backend/openapi-spec.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SkillBridge API Ecosystem",
    "description": "Skill Intelligence and Career Readiness Platform API Specifications",
    "contact": {
      "name": "SkillBridge Development Team",
      "email": "support@skillbridge.internal"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://springdoc.org"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
    "/api/projects/{projectId}/progress": {
      "put": {
        "tags": [
          "Project Service"
        ],
        "summary": "Update project completion progress",
        "operationId": "updateProgress",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectProgressRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseProjectProgress"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/projects/{projectId}/submit": {
      "post": {
        "tags": [
          "Project Service"
        ],
        "summary": "Submit GitHub repository URL for project evaluation",
        "operationId": "submitProject",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectSubmissionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseProjectProgress"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{projectId}/start": {
      "post": {
        "tags": [
          "Project Service"
        ],
        "summary": "Start working on a project",
        "operationId": "startProject",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseProjectProgress"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/learning/{courseId}/progress": {
      "post": {
        "tags": [
          "Learning Recommendation Service"
        ],
        "summary": "Update learning progress for a course",
        "operationId": "updateProgress_1",
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProgressUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseLearningProgress"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/skills": {
      "get": {
        "tags": [
          "Skill Service"
        ],
        "summary": "Get all available master skills",
        "operationId": "getSkills",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListSkill"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/readiness/role/{roleId}": {
      "get": {
        "tags": [
          "Career Readiness & Skill Gap Engine"
        ],
        "summary": "Get career readiness analysis for a specific role",
        "operationId": "getReadinessForRole",
        "parameters": [
          {
            "name": "roleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseCareerReadinessDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/projects/{projectId}": {
      "get": {
        "tags": [
          "Project Service"
        ],
        "summary": "Get project details by ID",
        "operationId": "getProjectById",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseProject"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{projectId}/evaluation": {
      "get": {
        "tags": [
          "Project Service"
        ],
        "summary": "Get project evaluation feedback",
        "operationId": "getEvaluation",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseProjectProgress"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ProjectProgressRequest": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "NOT_STARTED",
              "IN_PROGRESS",
              "SUBMITTED",
              "EVALUATED"
            ]
          },
          "progressPercent": {
            "type": "integer",
            "format": "int32"
          },
          "githubRepoUrl": {
            "type": "string"
          }
        }
      },
      "ApiResponseProjectProgress": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/ProjectProgress"
          },
          "message": {
            "type": "string"
          },
          "errorCode": {
            "type": "string"
          }
        }
      },
      "ProjectProgress": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "projectId": {
            "type": "integer",
            "format": "int64"
          },
          "status": {
            "type": "string",
            "enum": [
              "NOT_STARTED",
              "IN_PROGRESS",
              "SUBMITTED",
              "EVALUATED"
            ]
          },
          "githubRepoUrl": {
            "type": "string"
          },
          "progressPercent": {
            "type": "integer",
            "format": "int32"
          },
          "submissionNotes": {
            "type": "string"
          },
          "evaluationScore": {
            "type": "integer",
            "format": "int32"
          },
          "feedback": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "StudentSkill": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "skillId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "level": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "VERIFIED",
              "PARTIAL",
              "MISSING",
              "CLAIMED"
            ]
          },
          "score": {
            "type": "integer",
            "format": "int32"
          },
          "evidenceCount": {
            "type": "integer",
            "format": "int32"
          },
          "evidenceListJson": {
            "type": "string"
          }
        }
      },
...
      "ProjectSubmissionRequest": {
        "required": [
          "githubRepoUrl"
        ],
        "type": "object",
        "properties": {
          "githubRepoUrl": {
            "type": "string"
          },
          "submissionNotes": {
            "type": "string"
          }
        }
      },
...
      "ProgressUpdateRequest": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "NOT_STARTED",
              "IN_PROGRESS",
              "COMPLETED"
            ]
          },
          "progressPercent": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "ApiResponseLearningProgress": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/LearningProgress"
          },
          "message": {
            "type": "string"
          },
          "errorCode": {
            "type": "string"
          }
        }
      },
      "LearningProgress": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "courseId": {
            "type": "integer",
            "format": "int64"
          },
          "status": {
            "type": "string",
            "enum": [
              "NOT_STARTED",
              "IN_PROGRESS",
              "COMPLETED"
            ]
          },
          "progressPercent": {
            "type": "integer",
            "format": "int32"
          },
          "lastAccessedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ApiResponseListSkill": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Skill"
            }
          },
          "message": {
            "type": "string"
          },
          "errorCode": {
            "type": "string"
          }
        }
      },
      "Skill": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "skillId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
...
      "ApiResponseCareerReadinessDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/CareerReadinessDto"
          },
          "message": {
            "type": "string"
          },
          "errorCode": {
            "type": "string"
          }
        }
      },
      "CareerReadinessDto": {
        "type": "object",
        "properties": {
          "targetRoleId": {
            "type": "string"
          },
          "targetRoleTitle": {
            "type": "string"
          },
          "competencyCoveragePercentage": {
            "type": "integer",
            "format": "int32"
          },
          "calculationFormulaExplanation": {
            "type": "string"
          },
          "verifiedSkills": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StudentSkill"
            }
          },
          "partialSkills": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StudentSkill"
            }
          },
          "missingSkills": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StudentSkill"
            }
          },
          "skillGaps": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SkillGapDto"
            }
          },
          "nextBestAction": {
            "$ref": "#/components/schemas/NextActionDto"
          },
          "learningRecommendations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Course"
            }
          },
          "projectRecommendations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Project"
            }
          }
        }
      },
      "Course": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "platform": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "duration": {
            "type": "string"
          },
          "difficulty": {
            "type": "string"
          },
          "skillsCovered": {
            "type": "string"
          },
          "competenciesCovered": {
            "type": "string"
          },
          "isFree": {
            "type": "boolean"
          },
          "recommendationReason": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "skillId": {
            "type": "string"
          }
        }
      },
      "NextActionDto": {
        "type": "object",
        "properties": {
          "actionType": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "skillId": {
            "type": "string"
          },
          "assessmentId": {
            "type": "integer",
            "format": "int64"
          },
          "courseId": {
            "type": "integer",
            "format": "int64"
          },
          "projectId": {
            "type": "integer",
            "format": "int64"
          },
          "priority": {
            "type": "string"
          }
        }
      },
      "Project": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "objective": {
            "type": "string"
          },
          "skillsCovered": {
            "type": "string"
          },
          "requirements": {
            "type": "string"
          },
          "architecture": {
            "type": "string"
          },
          "estimatedTime": {
            "type": "string"
          },
          "difficulty": {
            "type": "string"
          },
          "evaluationCriteria": {
            "type": "string"
          },
          "submissionRequirements": {
            "type": "string"
          }
        }
      },
      "SkillGapDto": {
        "type": "object",
        "properties": {
          "skillId": {
            "type": "string"
          },
          "skillName": {
            "type": "string"
          },
          "currentStatus": {
            "type": "string",
            "enum": [
              "VERIFIED",
              "PARTIAL",
              "MISSING",
              "CLAIMED"
            ]
          },
          "requiredLevel": {
            "type": "string"
          },
          "importanceWeight": {
            "type": "number",
            "format": "double"
          },
          "gapPriority": {
            "type": "string",
            "enum": [
              "CRITICAL",
              "HIGH",
              "MEDIUM",
              "LOW"
            ]
          },
          "recommendationReason": {
            "type": "string"
          }
        }
      },
...
      "ApiResponseProject": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/Project"
          },
          "message": {
            "type": "string"
          },
          "errorCode": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "name": "bearerAuth",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}